+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fix Storage Emulator crashing with NoClassDefFoundError in some cases (#3481).
+- Refactor mechanism for invoking function triggers (#4886).